WIP: Support for Git dependencies
Git dependencies not only need to be fetched from the specified remote
repository, but they also need to have their checksums computed and
stored in a checksums file, as well as have all their packages unnested
and placed in flat structure.

This commit implements the former, but still work to cover the latter
still needs to be done.

Signed-off-by: Bruno Pimentel <bpimente@redhat.com>