Building from source: Install createrepo_c from a Red Hat subscription
Since building createrepo_c from source has proven to be very
troublesome, the alternative is installing it directly from an RPM.

The createrepo_c RPM is available from the Red Hat Satellite 6.16 for
RHEL 9 repository, which is only accessible with a Red Hat subscription.

For this reason, we need to change the rpms.lock.yaml file to rely on
the newly added redhat.repo file. An activation key is also necessary in
order to prefetch the content from this repository, and it will be
configured in Konflux.

Signed-off-by: Bruno Pimentel <bpimente@redhat.com>