Building from source: move the basic Dockerfile to Fedora
By removing the createrepo_c pip package, we now need to install it from
an RPM, which UBI only provides as entitled content. Moving the base
image to Fedora allows us to install this RPM directly.

This change also drops the dependency on Red Hat's subscription manager,
since it is not needed in order for Hermeto's to work properly.

Signed-off-by: Bruno Pimentel <bpimente@redhat.com>