Konflux onboarding: add Renovate configuration
The newly introduced renovate.json file essentially copies the
configuration that currently exists in the dependabot.yml file.

The only configuration not carried over is the one that ignores updates
to pydantic-core, since we want to test Renovate's behavior regarding
this package, to avoid needing to convert the dependabot-pipcompile.yml
hook.

Signed-off-by: Bruno Pimentel <bpimente@redhat.com>